requirements.txt
flask
python-dotenv
pandas
requests
pytz
beautifulsoup4
supabase